+++ sig/yard2steep/yard2steep/parser.rbi
@@ -7,10 +7,6 @@ class Yard2steep::Parser
   def parse_defs: (Array<any>) -> void
   def parse_def: (Array<any>) -> void
   def parse_method_impl: (String, Integer, Array<any>) -> void
-  def extract_p_types!: (Integer) -> void
-  def parse_comment!: (String) -> void
-  def try_param_comment: (String) -> bool
-  def try_return_comment: (String) -> bool
   def parse_params: (Array<any>) -> Array<AST::PNode>
   def parse_paren_params: (Array<any>) -> Array<AST::PNode>
   def parse_no_paren_params: (Array<any>) -> Array<AST::PNode>
@@ -22,16 +18,9 @@ class Yard2steep::Parser
   def parse_command_args_add_block: (Array<any>) -> Array<String>
   def parse_method_add_arg: (Array<any>) -> void
   def parse_attr_reader: (any) -> void
-  def extract_comments: (String) -> Hash<String, String>
   def debug_print!: (String) -> void
   def type_node: (String) -> AST::PTypeNode
   def block_type_node: (String) -> AST::PTypeNode
-  def normalize_type: (String) -> String
 end
-Yard2steep::Parser::S_RE: any
-Yard2steep::Parser::TYPE_WITH_PAREN_RE: any
-Yard2steep::Parser::COMMENT_RE: any
-Yard2steep::Parser::PARAM_RE: any
-Yard2steep::Parser::RETURN_RE: any
 Yard2steep::Parser::ANY_TYPE: any
 Yard2steep::Parser::ANY_BLOCK_TYPE: any
